All work load
We first tried to build api with FastAPI but we faced many challenges that we had tried to resolve till the end without issue.
Then we build the app on the other hand with Flask and it worked well on last minutes. We also try to deploy with Dockerfile on Azure but the building docker image did not work

Co-Authored-By: cmamakan <134331266+cmamakan@users.noreply.github.com>